refactor: (모달) 타입, 스타일, 훅, context 파일 별도 분리
